upgrade-handler code gen: investigating why e2e does not pass (#4080)
* updates

* fix
