+++ app/upgrades/v14/constants.go
@@ -6,11 +6,7 @@ import (
 	ibchookstypes "github.com/osmosis-labs/osmosis/x/ibc-hooks/types"
 
 	"github.com/osmosis-labs/osmosis/v14/app/upgrades"
-	cltypes "github.com/osmosis-labs/osmosis/v14/x/concentrated-liquidity/types"
 	downtimetypes "github.com/osmosis-labs/osmosis/v14/x/downtime-detector/types"
-	poolmanagertypes "github.com/osmosis-labs/osmosis/v14/x/poolmanager/types"
-	protorevtypes "github.com/osmosis-labs/osmosis/v14/x/protorev/types"
-	valsetpreftypes "github.com/osmosis-labs/osmosis/v14/x/valset-pref/types"
 )
 
 // UpgradeName defines the on-chain upgrade name for the Osmosis v14 upgrade.
@@ -20,7 +16,7 @@ var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
 	StoreUpgrades: store.StoreUpgrades{
-		Added:   []string{valsetpreftypes.StoreKey, protorevtypes.StoreKey, poolmanagertypes.StoreKey, downtimetypes.StoreKey, ibchookstypes.StoreKey, cltypes.StoreKey},
+		Added:   []string{downtimetypes.StoreKey, ibchookstypes.StoreKey},
 		Deleted: []string{},
 	},
 }
